refactor(postgres): split store orchestration modules
